Split first part of the job C bindings in their own files
