+++ src/unix/lwt_unix_unix.h
@@ -45,312 +45,23 @@
 /* +-----------------------------------------------------------------+
    | JOB: guess_blocking                                             |
    +-----------------------------------------------------------------+ */
-
-struct job_guess_blocking {
-    struct lwt_unix_job job;
-    int fd;
-    int result;
-};
-
-static void worker_guess_blocking(struct job_guess_blocking *job)
-{
-    struct stat stat;
-    if (fstat(job->fd, &stat) == 0)
-        job->result = !(S_ISFIFO(stat.st_mode) || S_ISSOCK(stat.st_mode));
-    else
-        job->result = 1;
-}
-
-static value result_guess_blocking(struct job_guess_blocking *job)
-{
-    value result = Val_bool(job->result);
-    lwt_unix_free_job(&job->job);
-    return result;
-}
-
-CAMLprim value lwt_unix_guess_blocking_job(value val_fd)
-{
-    LWT_UNIX_INIT_JOB(job, guess_blocking, 0);
-    job->fd = Int_val(val_fd);
-    return lwt_unix_alloc_job(&(job->job));
-}
-
 /* +-----------------------------------------------------------------+
    | JOB: wait_mincore                                               |
    +-----------------------------------------------------------------+ */
-
-#ifdef __CYGWIN__
-LWT_NOT_AVAILABLE2(unix_wait_mincore_job)
-#else
-struct job_wait_mincore {
-    struct lwt_unix_job job;
-    char *ptr;
-};
-
-static void worker_wait_mincore(struct job_wait_mincore *job)
-{
-    /* Read the byte to force the kernel to fetch the page: */
-    char dummy;
-    memcpy(&dummy, job->ptr, 1);
-}
-
-static value result_wait_mincore(struct job_wait_mincore *job)
-{
-    lwt_unix_free_job(&job->job);
-    return Val_unit;
-}
-
-CAMLprim value lwt_unix_wait_mincore_job(value val_buffer, value val_offset)
-{
-    LWT_UNIX_INIT_JOB(job, wait_mincore, 0);
-    job->ptr = (char *)Caml_ba_data_val(val_buffer) + Long_val(val_offset);
-    return lwt_unix_alloc_job(&(job->job));
-}
-#endif
-
 /* +-----------------------------------------------------------------+
    | JOB: open                                                       |
    +-----------------------------------------------------------------+ */
-
-#ifndef O_NONBLOCK
-#define O_NONBLOCK O_NDELAY
-#endif
-#ifndef O_DSYNC
-#define O_DSYNC 0
-#endif
-#ifndef O_SYNC
-#define O_SYNC 0
-#endif
-#ifndef O_RSYNC
-#define O_RSYNC 0
-#endif
-
-static int open_flag_table[] = {
-    O_RDONLY, O_WRONLY, O_RDWR,  O_NONBLOCK, O_APPEND, O_CREAT, O_TRUNC,
-    O_EXCL,   O_NOCTTY, O_DSYNC, O_SYNC,     O_RSYNC,  0, /* O_SHARE_DELETE,
-                                                             Windows-only */
-    0, /* O_CLOEXEC, treated specially */
-    0  /* O_KEEPEXEC, treated specially */
-};
-
-enum { CLOEXEC = 1, KEEPEXEC = 2 };
-
-static int open_cloexec_table[15] = {0, 0, 0, 0, 0, 0,       0,       0,
-                                     0, 0, 0, 0, 0, CLOEXEC, KEEPEXEC};
-
-struct job_open {
-    struct lwt_unix_job job;
-    int flags;
-    int perms;
-    int fd; /* will have value CLOEXEC or KEEPEXEC on entry to worker_open */
-    int blocking;
-    int error_code;
-    char *name;
-    char data[];
-};
-
-static void worker_open(struct job_open *job)
-{
-    int fd;
-    int cloexec;
-
-    if (job->fd & CLOEXEC)
-        cloexec = 1;
-    else if (job->fd & KEEPEXEC)
-        cloexec = 0;
-    else
-#if OCAML_VERSION_MAJOR >= 4 && OCAML_VERSION_MINOR >= 5
-        cloexec = unix_cloexec_default;
-#else
-        cloexec = 0;
-#endif
-
-#if defined(O_CLOEXEC)
-    if (cloexec) job->flags |= O_CLOEXEC;
-#endif
-
-    fd = open(job->name, job->flags, job->perms);
-#if !defined(O_CLOEXEC) && defined(FD_CLOEXEC)
-    if (fd >= 0 && cloexec) {
-        int flags = fcntl(fd, F_GETFD, 0);
-
-        if (flags == -1 || fcntl(fd, F_SETFD, flags | FD_CLOEXEC) == -1) {
-            int serrno = errno;
-            close(fd);
-            errno = serrno;
-            fd = -1;
-        }
-    }
-#endif
-    job->fd = fd;
-    job->error_code = errno;
-    if (fd >= 0) {
-        struct stat stat;
-        if (fstat(fd, &stat) < 0)
-            job->blocking = 1;
-        else
-            job->blocking = !(S_ISFIFO(stat.st_mode) || S_ISSOCK(stat.st_mode));
-    }
-}
-
-static value result_open(struct job_open *job)
-{
-    int fd = job->fd;
-    LWT_UNIX_CHECK_JOB_ARG(job, fd < 0, "open", job->name);
-    value result = caml_alloc_tuple(2);
-    Field(result, 0) = Val_int(fd);
-    Field(result, 1) = Val_bool(job->blocking);
-    lwt_unix_free_job(&job->job);
-    return result;
-}
-
-CAMLprim value lwt_unix_open_job(value name, value flags, value perms)
-{
-    LWT_UNIX_INIT_JOB_STRING(job, open, 0, name);
-    job->fd = caml_convert_flag_list(flags, open_cloexec_table);
-    job->flags = caml_convert_flag_list(flags, open_flag_table);
-    job->perms = Int_val(perms);
-    return lwt_unix_alloc_job(&(job->job));
-}
-
 /* +-----------------------------------------------------------------+
    | JOB: read                                                       |
    +-----------------------------------------------------------------+ */
-
-struct job_read {
-    struct lwt_unix_job job;
-    /* The file descriptor. */
-    int fd;
-    /* The amount of data to read. */
-    long length;
-    /* The OCaml string. */
-    value string;
-    /* The offset in the string. */
-    long offset;
-    /* The result of the read syscall. */
-    long result;
-    /* The value of errno. */
-    int error_code;
-    /* The temporary buffer. */
-    char buffer[];
-};
-
-static void worker_read(struct job_read *job)
-{
-    job->result = read(job->fd, job->buffer, job->length);
-    job->error_code = errno;
-}
-
-static value result_read(struct job_read *job)
-{
-    long result = job->result;
-    if (result < 0) {
-        int error_code = job->error_code;
-        caml_remove_generational_global_root(&(job->string));
-        lwt_unix_free_job(&job->job);
-        unix_error(error_code, "read", Nothing);
-    } else {
-        memcpy(String_val(job->string) + job->offset, job->buffer, result);
-        caml_remove_generational_global_root(&(job->string));
-        lwt_unix_free_job(&job->job);
-        return Val_long(result);
-    }
-}
-
-CAMLprim value lwt_unix_read_job(value val_fd, value val_buffer,
-                                 value val_offset, value val_length)
-{
-    long length = Long_val(val_length);
-    LWT_UNIX_INIT_JOB(job, read, length);
-    job->fd = Int_val(val_fd);
-    job->length = length;
-    job->string = val_buffer;
-    job->offset = Long_val(val_offset);
-    caml_register_generational_global_root(&(job->string));
-    return lwt_unix_alloc_job(&(job->job));
-}
-
 /* +-----------------------------------------------------------------+
    | JOB: bytes_read                                                 |
    +-----------------------------------------------------------------+ */
-
-struct job_bytes_read {
-    struct lwt_unix_job job;
-    /* The file descriptor. */
-    int fd;
-    /* The destination buffer. */
-    char *buffer;
-    /* The offset in the string. */
-    long offset;
-    /* The amount of data to read. */
-    long length;
-    /* The result of the read syscall. */
-    long result;
-    /* The value of errno. */
-    int error_code;
-};
-
-static void worker_bytes_read(struct job_bytes_read *job)
-{
-    job->result = read(job->fd, job->buffer, job->length);
-    job->error_code = errno;
-}
-
-static value result_bytes_read(struct job_bytes_read *job)
-{
-    long result = job->result;
-    LWT_UNIX_CHECK_JOB(job, result < 0, "read");
-    lwt_unix_free_job(&job->job);
-    return Val_long(result);
-}
-
-CAMLprim value lwt_unix_bytes_read_job(value val_fd, value val_buf,
-                                       value val_ofs, value val_len)
-{
-    LWT_UNIX_INIT_JOB(job, bytes_read, 0);
-    job->fd = Int_val(val_fd);
-    job->buffer = (char *)Caml_ba_data_val(val_buf) + Long_val(val_ofs);
-    job->length = Long_val(val_len);
-    return lwt_unix_alloc_job(&(job->job));
-}
-
 /* +-----------------------------------------------------------------+
    | JOB: write                                                      |
    +-----------------------------------------------------------------+ */
 
-struct job_write {
-    struct lwt_unix_job job;
-    int fd;
-    long length;
-    long result;
-    int error_code;
-    char buffer[];
-};
 
-static void worker_write(struct job_write *job)
-{
-    job->result = write(job->fd, job->buffer, job->length);
-    job->error_code = errno;
-}
-
-static value result_write(struct job_write *job)
-{
-    long result = job->result;
-    LWT_UNIX_CHECK_JOB(job, result < 0, "write");
-    lwt_unix_free_job(&job->job);
-    return Val_long(result);
-}
-
-CAMLprim value lwt_unix_write_job(value val_fd, value val_string,
-                                  value val_offset, value val_length)
-{
-    long length = Long_val(val_length);
-    LWT_UNIX_INIT_JOB(job, write, length);
-    job->fd = Int_val(val_fd);
-    job->length = length;
-    memcpy(job->buffer, String_val(val_string) + Long_val(val_offset), length);
-    return lwt_unix_alloc_job(&(job->job));
-}
 
 /* +-----------------------------------------------------------------+
    | JOB: bytes_write                                                |
