+++ src/unix/lwt_unix_windows.h
@@ -27,196 +27,16 @@ CAMLprim value lwt_unix_is_socket(value fd)
 }
 
 /* +-----------------------------------------------------------------+
-   | JOB: read                                                       |
-   +-----------------------------------------------------------------+ */
-
-struct job_read {
-    struct lwt_unix_job job;
-    union {
-        HANDLE handle;
-        SOCKET socket;
-    } fd;
-    int kind;
-    DWORD length;
-    DWORD result;
-    DWORD error_code;
-    value string;
-    DWORD offset;
-    char buffer[];
-};
-
-static void worker_read(struct job_read *job)
-{
-    if (job->kind == KIND_SOCKET) {
-        int ret;
-        ret = recv(job->fd.socket, job->buffer, job->length, 0);
-        if (ret == SOCKET_ERROR) job->error_code = WSAGetLastError();
-        job->result = ret;
-    } else {
-        if (!ReadFile(job->fd.handle, job->buffer, job->length, &(job->result),
-                      NULL))
-            job->error_code = GetLastError();
-    }
-}
-
-static value result_read(struct job_read *job)
-{
-    value result;
-    DWORD error = job->error_code;
-    if (error) {
-        caml_remove_generational_global_root(&job->string);
-        lwt_unix_free_job(&job->job);
-        win32_maperr(error);
-        uerror("read", Nothing);
-    }
-    memcpy(String_val(job->string) + job->offset, job->buffer, job->result);
-    result = Val_long(job->result);
-    caml_remove_generational_global_root(&job->string);
-    lwt_unix_free_job(&job->job);
-    return result;
-}
-
-CAMLprim value lwt_unix_read_job(value val_fd, value val_string,
-                                 value val_offset, value val_length)
-{
-    struct filedescr *fd = (struct filedescr *)Data_custom_val(val_fd);
-    long length = Long_val(val_length);
-    LWT_UNIX_INIT_JOB(job, read, length);
-    job->kind = fd->kind;
-    if (fd->kind == KIND_HANDLE)
-        job->fd.handle = fd->fd.handle;
-    else
-        job->fd.socket = fd->fd.socket;
-    job->length = length;
-    job->error_code = 0;
-    job->string = val_string;
-    job->offset = Long_val(val_offset);
-    caml_register_generational_global_root(&(job->string));
-    return lwt_unix_alloc_job(&(job->job));
-}
-
-/* +-----------------------------------------------------------------+
    | JOB: bytes_read                                                 |
    +-----------------------------------------------------------------+ */
 
-struct job_bytes_read {
-    struct lwt_unix_job job;
-    union {
-        HANDLE handle;
-        SOCKET socket;
-    } fd;
-    int kind;
-    char *buffer;
-    DWORD length;
-    DWORD result;
-    DWORD error_code;
-};
 
-static void worker_bytes_read(struct job_bytes_read *job)
-{
-    if (job->kind == KIND_SOCKET) {
-        int ret;
-        ret = recv(job->fd.socket, job->buffer, job->length, 0);
-        if (ret == SOCKET_ERROR) job->error_code = WSAGetLastError();
-        job->result = ret;
-    } else {
-        if (!ReadFile(job->fd.handle, job->buffer, job->length, &(job->result),
-                      NULL))
-            job->error_code = GetLastError();
-    }
-}
-
-static value result_bytes_read(struct job_bytes_read *job)
-{
-    value result;
-    DWORD error = job->error_code;
-    if (error) {
-        lwt_unix_free_job(&job->job);
-        win32_maperr(error);
-        uerror("bytes_read", Nothing);
-    }
-    result = Val_long(job->result);
-    lwt_unix_free_job(&job->job);
-    return result;
-}
-
-CAMLprim value lwt_unix_bytes_read_job(value val_fd, value val_buffer,
-                                       value val_offset, value val_length)
-{
-    struct filedescr *fd = (struct filedescr *)Data_custom_val(val_fd);
-    LWT_UNIX_INIT_JOB(job, bytes_read, 0);
-    job->kind = fd->kind;
-    if (fd->kind == KIND_HANDLE)
-        job->fd.handle = fd->fd.handle;
-    else
-        job->fd.socket = fd->fd.socket;
-    job->buffer = (char *)Caml_ba_data_val(val_buffer) + Long_val(val_offset);
-    job->length = Long_val(val_length);
-    job->error_code = 0;
-    return lwt_unix_alloc_job(&(job->job));
-}
 
 /* +-----------------------------------------------------------------+
    | JOB: write                                                      |
    +-----------------------------------------------------------------+ */
 
-struct job_write {
-    struct lwt_unix_job job;
-    union {
-        HANDLE handle;
-        SOCKET socket;
-    } fd;
-    int kind;
-    DWORD length;
-    DWORD result;
-    DWORD error_code;
-    char buffer[];
-};
 
-static void worker_write(struct job_write *job)
-{
-    if (job->kind == KIND_SOCKET) {
-        int ret;
-        ret = send(job->fd.socket, job->buffer, job->length, 0);
-        if (ret == SOCKET_ERROR) job->error_code = WSAGetLastError();
-        job->result = ret;
-    } else {
-        if (!WriteFile(job->fd.handle, job->buffer, job->length, &(job->result),
-                       NULL))
-            job->error_code = GetLastError();
-    }
-}
-
-static value result_write(struct job_write *job)
-{
-    value result;
-    DWORD error = job->error_code;
-    if (error) {
-        lwt_unix_free_job(&job->job);
-        win32_maperr(error);
-        uerror("write", Nothing);
-    }
-    result = Val_long(job->result);
-    lwt_unix_free_job(&job->job);
-    return result;
-}
-
-CAMLprim value lwt_unix_write_job(value val_fd, value val_string,
-                                  value val_offset, value val_length)
-{
-    struct filedescr *fd = (struct filedescr *)Data_custom_val(val_fd);
-    long length = Long_val(val_length);
-    LWT_UNIX_INIT_JOB(job, write, length);
-    job->kind = fd->kind;
-    if (fd->kind == KIND_HANDLE)
-        job->fd.handle = fd->fd.handle;
-    else
-        job->fd.socket = fd->fd.socket;
-    memcpy(job->buffer, String_val(val_string) + Long_val(val_offset), length);
-    job->length = length;
-    job->error_code = 0;
-    return lwt_unix_alloc_job(&(job->job));
-}
 
 /* +-----------------------------------------------------------------+
    | JOB: bytes_write                                                |
@@ -374,9 +194,6 @@ CAMLprim value lwt_unix_system_job(value cmdline)
    | Unavailable primitives                                          |
    +-----------------------------------------------------------------+ */
 
-LWT_NOT_AVAILABLE2(unix_wait_mincore_job)
-LWT_NOT_AVAILABLE1(unix_guess_blocking_job)
-LWT_NOT_AVAILABLE1(unix_open_job)
 LWT_NOT_AVAILABLE1(unix_stat_job)
 LWT_NOT_AVAILABLE1(unix_lstat_job)
 LWT_NOT_AVAILABLE1(unix_fstat_job)
