+++ src/unix/windows_c/not_available.c
@@ -56,4 +56,7 @@ LWT_NOT_AVAILABLE4(unix_wait4)
 LWT_NOT_AVAILABLE1(unix_get_cpu)
 LWT_NOT_AVAILABLE1(unix_get_affinity)
 LWT_NOT_AVAILABLE2(unix_set_affinity)
+LWT_NOT_AVAILABLE1(unix_guess_blocking_job)
+LWT_NOT_AVAILABLE2(unix_wait_mincore_job)
+LWT_NOT_AVAILABLE1(unix_open_job)
 #endif
